requirements.txt
torch
numpy
torchvision
lightly
matplotlib
Pillow
scikit-learn